JParkour v1.5 | Innovations, fixes
Added:
- Now you can create multiple map points and move the chicken anywhere!
- Setting up whether it is possible to receive / give damage to someone who is in line.
- Cooldown for restart (with a message)
- Accelerated HEX and correction of incorrect colors when using minimessage mode. + check for minimessage support to avoid errors.
- some updates that might not have been included here...

Fixed:
- The code has been optimized.
- Some names have been changed for the standard.
- Updates from some releases may have been unsuccessful. (Everything is fine now)
- The running of the essence (chicken) has become more productive.
- Greater division into classes.

Instructions:
- Updating from the first version to the latest.
If you are updating from the very first version to the new one, you may encounter that the stands and entity (chicken) movement do not match the previous values. You can restore all these values by going to the plugin editor, highlighting the positions, and re-saving the map, after which everything should return to its standard form.
- Some of the original items in the editor do not work after the update.
You can fix this by assigning the correct stands in slots.yml or by changing the name of the stands in config.yml and the schema file.

I am looking for developers/testers who can watch this plugin and check how it works for you. Many thanks in advance to all those who tested it and expressed their opinion about it, as it contains many corrections and additions.

The plugin shows the best stability on these versions: 1.8.8, 1.12.2, 1.16.5